requirements.txt
openai
tweepy
requests
apify-client
python-dotenv
